+++ packages/core/src/components/accordion/accordion.tsx
@@ -1,22 +1,6 @@
-import {
-  h,
-  Component,
-  ComponentInterface,
-  Prop,
-  Host,
-  Method,
-  Watch,
-  Element,
-  Event,
-  EventEmitter,
-} from "@stencil/core";
+import { h, Component, ComponentInterface, Prop, Host, Method, Watch, Element } from "@stencil/core";
 
-import {
-  AccordionInternalState,
-  AccordionSectionToggleAnimationEndEvent,
-  AccordionSectionToggleEvent,
-  AccordionVariant,
-} from "./accordion.interfaces";
+import { AccordionInternalState, AccordionVariant } from "./accordion.interfaces";
 
 import { createStore } from "@stencil/store";
 
@@ -35,58 +19,24 @@ export class Accordion implements ComponentInterface {
    * The variant of the Accordion.
    */
   @Prop({ reflect: true })
-  variant?: AccordionVariant = "default";
+  variant: AccordionVariant = "default";
 
   /**
-   * Places the chevron at the opposite side. Note: this mode does not display `state`, `attachmentCount` or `status` props on child `<dso-accordion-section>` elements
+   * Places the chevron at the opposite side.
+   *
+   * Note: this mode does not display `state`, `attachmentCount` or `status` props on Accordion Sections
    */
   @Prop({ reflect: true })
   reverseAlign = false;
 
-  /**
-   * Allows multiple sections to be open at the same time.
-   */
-  @Prop({ reflect: true })
-  allowMultipleOpen = false;
-
-  /**
-   * Emitted when a section is toggled.
-   *
-   * `event.detail.originalEvent` contains the original `MouseEvent` when the section is toggled by clicking on the header
-   * `event.detail.section` contains the toggled section and its new opened value.\
-   * `event.detail.sections` contains all `<dso-accordion-section>` elements belonging to this accordion.
-   */
-  @Event()
-  dsoToggleSection!: EventEmitter<AccordionSectionToggleEvent>;
-
-  /**
-   * Event emitted when the accordion section completes its toggle animation.
-   */
-  @Event()
-  dsoToggleSectionAnimationEnd!: EventEmitter<AccordionSectionToggleAnimationEndEvent>;
-
   @Watch("variant")
   updateVariant(variant: AccordionVariant = "default") {
-    this.accordionState.variant = variant || "default";
+    this.accordionState.variant = variant;
   }
 
   @Watch("reverseAlign")
   updateReverseAlign(reverseAlign: boolean) {
     this.accordionState.reverseAlign = reverseAlign;
-  }
-
-  @Watch("allowMultipleOpen")
-  updateAllowMultipleOpen(allowMultipleOpen: boolean) {
-    this.accordionState.allowMultipleOpen = allowMultipleOpen;
-
-    if (!allowMultipleOpen) {
-      const openSections = Array.from(this.host.querySelectorAll<HTMLElement>(":scope > dso-accordion-section[open]"));
-
-      // By removing the first section, it is kept open;
-      openSections.shift();
-
-      openSections.forEach((section) => this.controlOpenAttribute(section, false));
-    }
   }
 
   /**
@@ -97,131 +47,22 @@ export class Accordion implements ComponentInterface {
     return this.accordionState;
   }
 
-  /**
-   * Toggle a section. Pass the `<dso-accordion-section>` element or the index of the section.
-   * @param sectionElement The section element that needs to toggle
-   * @param event The event that the user triggered
-   * @returns The state of the section
-   */
-  @Method()
-  async toggleSection(sectionElement: HTMLElement | number, event?: MouseEvent): Promise<undefined | boolean> {
-    const sections = Array.from(this.host.querySelectorAll<HTMLElement>(":scope > dso-accordion-section"));
-
-    if (typeof sectionElement === "number") {
-      const section = sections[sectionElement];
-      if (section instanceof HTMLElement) {
-        sectionElement = section;
-      }
-    }
-
-    if (!(sectionElement instanceof HTMLElement) || !sections.includes(sectionElement)) {
-      return;
-    }
-
-    const sectionIsOpen = this.isSectionOpen(sectionElement);
-
-    if (this.allowMultipleOpen) {
-      this.controlOpenAttribute(sectionElement, !sectionIsOpen);
-      this.emitToggleEvent(sectionElement, sections, event);
-      return !sectionIsOpen;
-    }
-
-    if (sectionIsOpen) {
-      this.controlOpenAttribute(sectionElement, false);
-      this.emitToggleEvent(sectionElement, sections, event);
-
-      return false;
-    }
-
-    await this.closeOpenSections();
-
-    this.controlOpenAttribute(sectionElement, true);
-    this.emitToggleEvent(sectionElement, sections, event);
-
-    return true;
-  }
-
-  /**
-   * Emitted when the animation of opening or closing ends.
-   */
-  @Method()
-  async animationEnd(sectionElement: HTMLElement): Promise<void> {
-    this.dsoToggleSectionAnimationEnd.emit({
-      section: {
-        element: sectionElement,
-        open: this.isSectionOpen(sectionElement),
-      },
-    });
-  }
-
-  /**
-   * Closes all sections belonging to this accordion.
-   */
-  @Method()
-  async closeOpenSections(): Promise<void> {
-    const sections = Array.from(this.host.querySelectorAll<HTMLElement>(":scope > dso-accordion-section"));
-
-    const openSections = sections.filter((s) => this.isSectionOpen(s));
-    openSections.forEach((section) => this.controlOpenAttribute(section, false));
-  }
-
-  // These checks are needed for a React timing issue.
   componentWillLoad() {
-    if (this.accordionState.variant !== this.variant) {
-      this.accordionState.variant = this.variant || "default";
-    }
-
-    if (this.accordionState.reverseAlign !== this.reverseAlign) {
-      this.accordionState.reverseAlign = this.reverseAlign;
-    }
-
-    if (this.accordionState.allowMultipleOpen !== this.allowMultipleOpen) {
-      this.accordionState.allowMultipleOpen = this.allowMultipleOpen;
-    }
+    this.accordionState.variant = this.variant;
+    this.accordionState.reverseAlign = this.reverseAlign;
   }
 
   constructor() {
-    const { state } = createStore<AccordionInternalState>({
-      variant: this.variant || "default",
-      reverseAlign: this.reverseAlign,
-      allowMultipleOpen: this.allowMultipleOpen,
-    });
+    const { state } = createStore<AccordionInternalState>({});
 
     this.accordionState = state;
   }
 
   render() {
     return (
-      <Host
-        class={{
-          "dso-accordion": true,
-        }}
-      >
+      <Host class="dso-accordion">
         <slot></slot>
       </Host>
     );
-  }
-
-  private emitToggleEvent(sectionElement: HTMLElement, sections: HTMLElement[], e?: MouseEvent): void {
-    this.dsoToggleSection.emit({
-      originalEvent: e,
-      section: {
-        element: sectionElement,
-        open: this.isSectionOpen(sectionElement),
-      },
-      sections,
-    });
-  }
-
-  private isSectionOpen(sectionElement: HTMLElement): boolean {
-    return typeof sectionElement.getAttribute("open") === "string";
-  }
-
-  private controlOpenAttribute(sectionElement: HTMLElement, setAttribute: boolean): void {
-    if (setAttribute) {
-      sectionElement.setAttribute("open", "");
-    } else {
-      sectionElement.removeAttribute("open");
-    }
   }
 }