+++ packages/core/src/components/expandable/expandable.functions.ts
@@ -1,3 +1,0 @@
-export function isExpandable(element: Element | null): element is HTMLDsoExpandableElement {
-  return element instanceof Element && element.tagName === "DSO-EXPANDABLE";
-}