+++ packages/react/src/components/accordion/accordion.stories.ts
@@ -1,14 +1,7 @@
 import { storiesOfAccordion } from "dso-toolkit";
 import { storiesOf } from "@storybook/react";
 import { templateContainer } from "../../templates";
-import {
-  addonsSections,
-  alignmentSections,
-  allowMultipleOpenSections,
-  anchorSections,
-  basicSections,
-  subSections,
-} from "./accordion.content";
+import { addonsSections, alignmentSections, anchorSections, basicSections, subSections } from "./accordion.content";
 
 import readme from "./readme.md?raw";
 
@@ -27,7 +20,6 @@ storiesOfAccordion({
     neutralSections: basicSections,
     anchorSections,
     subSections: subSections(templates),
-    allowMultipleOpenSections,
     addonsSections,
     alignmentSections,
   }),